+++ lib/modules/navigation/navigation_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: navigation
-// Updated at: 2026-04-04 15:19:36
-// Commit: feat(navigation): add async navigation stream provider
+// Updated at: 2026-04-04 19:39:48
+// Commit: feat(navigation): integrate custom telemetry logger in navigation
 
 class navigationComponentState {
-  final String commitRef = '115';
-  final String lastMessage = 'feat(navigation): add async navigation stream provider';
-  final DateTime timestamp = DateTime.parse('2026-04-04T15:19:36+05:00');
+  final String commitRef = '125';
+  final String lastMessage = 'feat(navigation): integrate custom telemetry logger in navigation';
+  final DateTime timestamp = DateTime.parse('2026-04-04T19:39:48+05:00');
 }